Restructure V3 schema to more closely align with V2
- Add base schema reader with V2 and V3 extensions
- Update tests
