+++ src/main/java/io/xlate/edi/internal/schema/SchemaReader.java
@@ -1,0 +1,16 @@
+package io.xlate.edi.internal.schema;
+
+import java.util.Map;
+
+import io.xlate.edi.schema.EDISchemaException;
+import io.xlate.edi.schema.EDIType;
+
+public interface SchemaReader {
+
+    Map<String, EDIType> readTypes() throws EDISchemaException;
+
+    String getInterchangeName();
+
+    String getTransactionName();
+
+}
