+++ src/main/java/io/xlate/edi/schema/implementation/SegmentImplementation.java
@@ -24,6 +24,6 @@ public interface SegmentImplementation extends EDITypeImplementation<EDIComplexT
 
     Discriminator getDiscriminator();
 
-    List<EDITypeImplementation<? extends EDIType>> getSequence();
+    List<EDITypeImplementation<EDIType>> getSequence();
 
 }
